refactor: extract sweep loading into independent module
Extracted `load_sweep`, `load_evaluation_results`, and related models into `src/run/load.rs`. This breaks the structural cycle where `compare.rs` depended on `evaluate.rs` for `EvaluationResults` and `evaluate.rs` depended back on `compare.rs` for `load_sweep`. All paths updated to enforce the new acyclic dependency graph.
