+++ src/run/budget_fit.rs
@@ -14,7 +14,9 @@ use std::path::{Path, PathBuf};
 use serde::{Deserialize, Serialize};
 
 use crate::error::Error;
-use crate::run::compare::{load_evaluation_results, load_sweep};
+use crate::run::evaluate::load_evaluation_results;
+use crate::run::load::load_sweep;
+
 use crate::run::swebench::{
     InstanceResult, SWEEP_STATUS_COMPLETED, resolved_count as instance_resolved_count,
 };
