+++ src/run/cache_stats.rs
@@ -14,7 +14,7 @@ use crate::cost::{
     ANTHROPIC_CACHE_CREATION_MULTIPLIER, ANTHROPIC_CACHE_READ_MULTIPLIER, SONNET_INPUT_USD_PER_MTOK,
 };
 use crate::error::Error;
-use crate::run::compare::load_sweep;
+use crate::run::load::load_sweep;
 
 // ── public calculation helpers (unit-testable) ────────────────────────────────
