+++ src/run/grep.rs
@@ -10,7 +10,7 @@ use serde::{Deserialize, Serialize};
 use crate::artifact::{ArtifactKind, classify_json_value};
 use crate::error::Error;
 use crate::redaction::{Redactor, surface};
-use crate::run::compare::load_sweep;
+use crate::run::load::load_sweep;
 use crate::trajectory::Trajectory;
 
 #[derive(Debug, Clone)]
