+++ src/run/instance_history.rs
@@ -10,7 +10,7 @@ use serde::{Deserialize, Serialize};
 
 use crate::error::Error;
 use crate::redaction::{Redactor, surface};
-use crate::run::compare::load_sweep;
+use crate::run::load::load_sweep;
 use crate::run::swebench::{effective_runs, resolved_count as instance_resolved_count};
 
 // ── output format ─────────────────────────────────────────────────────────────
