+++ src/run/near_miss.rs
@@ -22,7 +22,7 @@ use serde::{Deserialize, Serialize};
 
 use crate::artifact::ArtifactKind;
 use crate::error::Error;
-use crate::run::compare::load_evaluation_results;
+use crate::run::evaluate::load_evaluation_results;
 
 // ── format ───────────────────────────────────────────────────────────────────
