+++ src/run/stagnation_report.rs
@@ -22,7 +22,7 @@ use serde::{Deserialize, Serialize};
 use crate::artifact::ArtifactKind;
 use crate::error::Error;
 use crate::redaction::{Redactor, surface};
-use crate::run::compare::load_sweep;
+use crate::run::load::load_sweep;
 use crate::stagnation::{action_hash, canonicalize_action};
 use crate::trajectory::{FailureCategory, Trajectory};
