+++ src/run/triage.rs
@@ -10,7 +10,9 @@ use sha2::{Digest as _, Sha256};
 use crate::artifact::{ArtifactKind, classify_json_value};
 use crate::env::RunResult;
 use crate::error::Error;
-use crate::run::compare::{load_evaluation_results_checked, load_sweep};
+use crate::run::evaluate::load_evaluation_results_checked;
+use crate::run::load::load_sweep;
+
 use crate::run::swebench::{InstanceResult, resolved_count};
 use crate::trajectory::{FailureCategory, Trajectory};
