+++ src/run/triage_diff.rs
@@ -13,7 +13,9 @@ use std::path::PathBuf;
 use serde::{Deserialize, Serialize};
 
 use crate::error::Error;
-use crate::run::compare::{load_evaluation_results_checked, load_sweep};
+use crate::run::evaluate::load_evaluation_results_checked;
+use crate::run::load::load_sweep;
+
 use crate::run::triage::{
     TriageArgs, TriageCluster, TriageReport, candidate_instance_ids, extract_instance_signature,
     resolve_trajectory_path,
